refactor(dashboard): keep the track record copy inside the panel feature
The cases, testimonials and their links are panel copy — the track record
section and the testimonial carousel are the only readers, and nothing outside
the feature has a reason to become one. Sitting in shared/constants it invited
that dependency; under features/panel it stays where the section that owns it
lives.
